Correct exception specification
It's a happy coincidence that the documentation build caught this problem. :+1: 

The code to adapt to both `monotonic_ns` and `monotonic` is structured a bit differently than the LED animation library, which is not a problem.
However, when monotonic_ns is not available, the line `time.monotonic_ns()` creates an AttributeError, so we need to catch that.  Since there's no `import` inside this block, we do _NOT_ need to catch `ImportError`.  At any rate, that's my quick analysis of what has occurred here.  I did not do any testing.

I think that if I'm correct it will fix the documentation building problem, as well as the functionality.